[Fix] #20 - Refactor code
