+++ src/utils/GetEmotionIcon.tsx
